contact form added

+++ src/app/api/category/[slug]/route.js
@@ -1,10 +1,16 @@
 import connect from "@/config/dbConfig";
+import authOptions from "@/helpers/nextAuthOptions";
 import Category from "@/models/CategoryModel";
+import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 
 export const DELETE = async (request, { params }) => {
     await connect();
     try {
+        const session = await getServerSession(authOptions);
+        if (!session) {
+            return NextResponse.json({ message: 'Unauthorized access', success: false }, { status: 401 })
+        }
         const category = await Category.findOneAndDelete({ slug: params?.slug })
         return NextResponse.json({ message: 'Category deleted successfully', success: true, category }, { status: 200 })
     } catch (error) {
@@ -16,6 +22,10 @@ export const DELETE = async (request, { params }) => {
 export const PUT = async (request, { params }) => {
     await connect();
     try {
+        const session = await getServerSession(authOptions);
+        if (!session) {
+            return NextResponse.json({ message: 'Unauthorized access', success: false }, { status: 401 })
+        }
         const { name, slug } = await request.json();
         const category = await Category.findOneAndUpdate({ slug: params?.slug }, { name, slug }, { new: true })
         return NextResponse.json({ message: 'Category updated successfully', success: true, category }, { status: 200 })
